完善网关

+++ gate/constant/constant.go
@@ -10,4 +10,4 @@ const (
 
 func ConnName(connID uint64) string { return fmt.Sprintf("conn-%d", connID) }
 
-func StreamName(serverID uint64) string { return fmt.Sprintf("stream-%v", serverID) }
+func StreamName(serverID int32) string { return fmt.Sprintf("stream-%v", serverID) }
